+++ cropclassification/postprocess/classification_postprocess.py
@@ -116,7 +116,6 @@ def calc_top3_and_consolidation(
         # First add some aditional columns specific for the export
         pred_output_df = pred_df.copy()
         pred_output_df["markercode"] = conf.marker["markertype"]
-        pred_output_df["run_id"] = conf.general["run_id"]
         today = datetime.date.today()
         pred_output_df["cons_date"] = today
         pred_output_df["modify_date"] = today
@@ -131,7 +130,7 @@ def calc_top3_and_consolidation(
         if conf.marker["markertype"] in ["LANDCOVER", "LANDCOVER-EARLY"]:
             table_name = "mon_marker_landcover"
             table_columns = (
-                "layer_id, prc_id, versienummer, markercode, run_id, cons_input, "
+                "layer_id, prc_id, versienummer, markercode, cons_input, "
                 "cons_landcover, cons_status, cons_date date 'yyyy-mm-dd', landcover1, "
                 "probability1, landcover2, probability2, landcover3, probability3, "
                 "modify_date date 'yyyy-mm-dd'"
@@ -139,7 +138,7 @@ def calc_top3_and_consolidation(
         elif conf.marker["markertype"] in ["CROPGROUP", "CROPGROUP-EARLY"]:
             table_name = "mon_marker_cropgroup"
             table_columns = (
-                "layer_id, prc_id, versienummer, markercode, run_id, cons_input, "
+                "layer_id, prc_id, versienummer, markercode, cons_input, "
                 "cons_cropgroup, cons_status, cons_date date 'yyyy-mm-dd', cropgroup1, "
                 "probability1, cropgroup2, probability2, cropgroup3, probability3, "
                 "modify_date date 'yyyy-mm-dd'"